# backend/openapi_full.yaml
openapi: 3.0.0
info:
  title: BidRoom API
  version: 1.0.0
  description: Complete API documentation for BidRoom backend.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /admin/transactions:
    get:
      tags:
        - admin
      summary: Auto-generated summary for GET /admin/transactions
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
...
  /finance/transactions:
    get:
      tags:
        - finance
      summary: Auto-generated summary for GET /finance/transactions
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
    post:
      tags:
        - finance
      summary: Auto-generated summary for POST /finance/transactions
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                exampleField:
                  type: string
...
  /transactions/:
    post:
      tags:
        - transaction
      summary: Auto-generated summary for POST /transactions/
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                exampleField:
                  type: string
    get:
      tags:
        - transaction
      summary: Auto-generated summary for GET /transactions/
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
  /transactions/stats:
    get:
      tags:
        - transaction
      summary: Auto-generated summary for GET /transactions/stats
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
  /transactions/{id}:
    get:
      tags:
        - transaction
      summary: Auto-generated summary for GET /transactions/{id}
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
  /transactions/{id}/status:
    put:
      tags:
        - transaction
      summary: Auto-generated summary for PUT /transactions/{id}/status
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                exampleField:
                  type: string
  /transactions/{id}/refund:
    post:
      tags:
        - transaction
      summary: Auto-generated summary for POST /transactions/{id}/refund
      responses:
        '200':
          description: Successful response
        '401':
          description: Unauthorized
        '500':
          description: Server Error
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                exampleField:
                  type: string
...